net/dns, util/publicdns: extract public DNS mapping into own package
This extracts DOH mapping of known public DNS providers in
forwarder.go into its own package, to be consumed by other repos
